refactor: move predicates into their own folder under `utils`
