+++ src/rules/no-redundant-files.ts
@@ -3,7 +3,10 @@ import type * as ESTree from 'estree';
 import type { AST } from 'jsonc-eslint-parser';
 
 import { createRule } from '../createRule.ts';
-import { isJSONStringLiteral, isNotNullish } from '../utils/predicates.ts';
+import {
+  isJSONStringLiteral,
+  isNotNullish,
+} from '../utils/predicates/index.ts';
 
 const defaultFiles = [
   /* cspell:disable-next-line */
