+++ src/rules/restrict-top-level-properties.ts
@@ -2,7 +2,7 @@ import { fixRemoveObjectProperty, type ObjectProperty } from 'eslint-fix-utils';
 import type { AST } from 'jsonc-eslint-parser';
 
 import { createRule } from '../createRule.ts';
-import { isJSONStringLiteral } from '../utils/predicates.ts';
+import { isJSONStringLiteral } from '../utils/predicates/index.ts';
 
 export const rule = createRule({
   create(context) {
