+++ src/rules/specify-peers-locally.ts
@@ -1,7 +1,7 @@
 import type { AST } from 'jsonc-eslint-parser';
 
 import { createRule } from '../createRule.ts';
-import { isJSONStringLiteral } from '../utils/predicates.ts';
+import { isJSONStringLiteral } from '../utils/predicates/index.ts';
 
 export const rule = createRule({
   create(context) {
