+++ src/rules/unique-dependencies.ts
@@ -6,7 +6,10 @@ import type * as ESTree from 'estree';
 import type { AST } from 'jsonc-eslint-parser';
 
 import { createRule } from '../createRule.ts';
-import { isJSONStringLiteral, isNotNullish } from '../utils/predicates.ts';
+import {
+  isJSONStringLiteral,
+  isNotNullish,
+} from '../utils/predicates/index.ts';
 
 const dependencyPropertyNames = new Set([
   'bundledDependencies',
