+++ src/rules/valid-peerDependenciesMeta-relationship.ts
@@ -3,7 +3,7 @@ import type * as ESTree from 'estree';
 import type { AST } from 'jsonc-eslint-parser';
 
 import { createRule } from '../createRule.ts';
-import { isJSONStringLiteral } from '../utils/predicates.ts';
+import { isJSONStringLiteral } from '../utils/predicates/index.ts';
 
 export const rule = createRule({
   create(context) {
