+++ src/utils/createSimpleRequirePropertyRule.ts
@@ -1,7 +1,7 @@
 import type { AST } from 'jsonc-eslint-parser';
 
 import { createRule } from '../createRule.ts';
-import { isJSONStringLiteral } from './predicates.ts';
+import { isJSONStringLiteral } from './predicates/index.ts';
 
 export interface CreateRequirePropertyRuleOptions {
   /**
