+++ src/utils/predicates/index.ts
@@ -1,0 +1,3 @@
+export { isJSONStringLiteral } from './isJSONStringLiteral.ts';
+export { isNotNullish } from './isNotNullish.ts';
+export { isPackageJson } from './isPackageJson.ts';
