+++ src/utils/predicates/isNotNullish.ts
@@ -1,0 +1,5 @@
+export function isNotNullish<T extends NonNullable<unknown>>(
+  value: null | T | undefined,
+): value is T {
+  return value !== null && value !== undefined;
+}
