+++ src/utils/predicates/isPackageJson.test.ts
