+++ src/utils/predicates/isPackageJson.ts
